# Cache generated grammar regexps to improve performance