--- a/lib/Babble/Grammar.pm
+++ b/lib/Babble/Grammar.pm
@@ -15,6 +15,8 @@ lazy rules => sub {
   +{ map +($_ => [ undef ]), keys %{ $_[0]->base_rule_names } }
 };
 
+# global cache of compiled grammar regexps
+my %COMPILE_CACHE;
 lazy grammar_regexp => sub {
   my ($self) = @_;
   my @parts;
@@ -35,7 +37,7 @@ lazy grammar_regexp => sub {
   # This stringify is required for Perl v5.18 - v5.28
   # (RT #126285, RT #144248).
   my $final_re = "${define_block} ${base_re}";
-  return qr{$final_re}x;
+  return $COMPILE_CACHE{$final_re} ||= qr{$final_re}x;
 };
 
 sub _rule_name {